Add files via upload
This version contains most of the code that will be used on front end. The tree structure is not complete, but everything else were tested with constants. More tests will be applied after I got the other classes. It's really hard to test a driver without the actual backend.

Justin